Simplifying macro 2.10 compat.
